ParrinelloRahmanBarostat builder
This is part of task #3437, implementing builders to create the
modules of modular simulator. The approach is explained in
docs/doxygen/lib/modularsimulator.md, "### Building modules".

The current change implements a builder for the Parrinello-Rahman
barostat. This also fixes a typo in a function of the energy element,
removes some unnecessary `gmx::` namespace specifiers, and uses
list initializer for vectors and matrices (which was previously throwing
off uncrustify).

Refs #3437