Propagator builder
This is part of task #3437, implementing builders to create the
modules of modular simulator. The approach is explained in
docs/doxygen/lib/modularsimulator.md, "### Building modules".

The current change implements a builder for the different propagators.
It also removes an obsolete `\todo` tag in the doxygen, and uses
list initialization in the propagator (which used to throw off uncrustify).

Refs #3437