Do not pass DeviceStream to look-up table creation function
The stream was only needed because cuda implementation relied on the
copyToDeviceBuffer(...) function, which is a general copy function.
Lookup tables are created once at the initialization stage and the
copy is made synchonously, hence nullptr stream can be used. This
makes the function siugnature shorter and the fuction - easier to
use.

Change-Id: I2c609c440048173103539930088ebed3758c9b7c