StatePropagatorData builder
This is part of task #3437, implementing builders to create the
modules of modular simulator. The approach is explained in
docs/doxygen/lib/modularsimulator.md, "### Building modules".

The current change implements a builder for the state propagator
data object. It also updates the constructor of StatePropagatorData
to use list initialization in the propagator (which used to throw
off uncrustify).

Refs #3437