Signaler builders
This is part of task #3437, implementing builders to create the
modules of modular simulator. The approach is explained in
docs/doxygen/lib/modularsimulator.md, "### Building modules".

The current change implements the new builder approach detailed
in the manual mentioned above to the signalers and the trajectory
element. These classes had a builder object already, but required
the elements to be built in a fixed order nonetheless. This change
also initializes an unitialized Step counter in the constructor of
TrajectoryElement, and cleans up some doxygen strings.

Refs #3437