Allow thermostat to scale velocities before and after step
Velocity scaling thermostats are currently only able to scale
velocities before the respective propagator step. This change allows
to also scale velocities at the end of the step. This functionality
is needed by the Nose-Hoover thermostat.

This also adds a few `Count` entries for class enums in propagator.h
and sorts the forward declarations in modularsimulatorinterfaces.h
alphabetically.

Refs #3423