Clean up simulator algorithm
The FreeEnergyData::Element was created instead of using the one
owned by the FreeEnergyData object which ensures them to have
identical life time. This didn't result in wrong execution, but
removed some of the life time safety which was supposed to be in
place.

The setup and teardown were run from the ownership list rather
than the dedicated setup / teardown list. As the two were identical
until now, this was not a problem, but should be avoided when having
elements which are not in the ownership list (but, for example,
owned by their respective data object...).

Finally, the fact that elements were stored in the ownership list,
but added to setup / teardown list and other infrastructure later
required work-arounds to ensure that no double counting was done.
This is simplified by doing the infrastructure work when elements
are stored.