Introduce ReferenceTemperatureManager
Simulated annealing and simulated tempering need to change the
reference temperature. As we want to keep the inputrec constant, we
are introducing a reference temperature manager. This object allows
elements to change the reference temperature, and it allows elements
to register callbacks so they are informed when the reference
temperature is changed.

The ReferenceTemperatureManager also changes the inputrec for
backwards compatibility with legacy functionality.

The current change introduces the ReferenceTemperatureManager, and
writes the callbacks for elements that need to be informed about
temperature changes. Simulated annealing and simulated tempering will
be introduced in subsequent commits.