Introduce Trotter decomposition / MTTK for modular simulator
Introduce the Trotter decomposition / MTTK temperature and pressure
coupling scheme into modular simulator. A few notes:

* The implementation uses the generic velocity and position scaling
propagators to avoid creating additional specialized types of combined
propagation and scaling elements.
* The MTTK pressure coupling and the Nose-Hoover chain temperature
coupling are implemented from scratch, as the current implementation
will be superseded by the modular simulator and removed as soon as the
modular simulator is feature complete.
* Since the Trotter decomposition requires the extended system
variables to be propagated by multiple instances of elements every
step, they are stored in objects whose lifetime are managed by the
simulator algorithm and can be accessed by the different elements
(without polluting the general simulation state as in the legacy
implementation).
* Since the element order in the Trotter-based algorithms are
significantly different from the other algorithms (leap-frog and
non-Trotter velocity verlet), it's creation is kept separate.

This also fixes a GMX_RELEASE_ASSERT in the templated
ModularSimulatorAlgorithmBuilder::storeSimulationData
and ModularSimulatorAlgorithmBuilder::simulationData functions which
failed to compile due to a string type mismatch.

Refs #3417, #3423