Add assertions on proper device is active for the requested stream in CUDA
In CUDA, the device is set active by a separate API call. It is important
to keep track on which device is currently set active, since the operation
can be requested in a stream that is attached to the not-active device. This
adds the assertions on whether the device context is active in CUDA API wrappers.
In places, where the performance is not important, release- level asserts are used.

Refs. #3805