Fix coloumb_tab uninitialized bug
nbp->coulomb_tab is being freed in init_ewald_coulomb_force_table()
whenever it is non-null. But it was not being initialized at startup,
causing an intermittent error in freeDeviceBuffer whenever it (by
chance) aquired a non-null value. This fixes the issue (which was
first observed after c4a672b93b), by initializing to a nullptr.