Move SIMD update to own header
This puts all update code with a SIMD component in its own header.
Also replaces some defines with if(constexpr). This also means that
a few extra, previously implicit, else branches needed to be added.

This is part of broader cleanup with the ultimate goal of having
the update director code live in update.cpp and the actual
algorithm implementations live in their own header/source files.