Add vector of vectors to t_state
This implements the possibility to add abtrirary RVec vectors
for atoms to t_state at runtime.
The cg_p vector is state now uses this new mechanism.

This change required some refactoring in minimize.cpp, as t_state
can no longer be copied. Now the vector contents are copied, which is
anyhow a better solution.