Refactor dev flags for CUDA-aware GPU features
Used mdlog.warning only when an environment variable was given and the
corresponding feature is not detected or not being activated.

Used MessageCollector where multiple possible reasons exist for a
decision.

Removed some log output that duplicated the report of CUDA awareness
of the MPI library in binaryinformation.cpp.

Also removed some cruft that suited gcc 5, which is no longer
supported.

Fixed logic about use of GPU update with FAHCORE

Fixes #4021