Unpack most t_lambda params to dispatchFreeEnergyKernel
Added a function that checks if soft core potentials are used and
pass all t_lambda data except all_lamda to dispatchFreeEnergyKernel.

Passing all_lambda is currently not feasible as EnumerationArray is
not a suitable type for passing through and interface due to the
need to include the EnumerationArray header. Further refactoring
will be needed before all_lambda could also be passed.