WIP: Fixed GPU force reduction event dependency bug
StatePropagatorDataGpu::Impl::getForcesReadyOnDeviceEvent() did not
correctly distinguish between the following cases:
- completion of the CPU force contribution H2D for GPU reduction
- force reduction completion
As a result, it always returned the event associated for the latter,
creating an incorrect dependency for the force reduction (which could
start before H2D copy was completed). This also resulted in a circular
dependency but as this is valid in CUDA it did not cause issues.

This commit changes getForcesReadyOnDeviceEvent() to never return
reduction-related events separating concerns; as a side-effect some
added complexity is required when invoking the GPU integrator.

TODO:
[ ] needs verification
[ ] release notes

Fixes #4032