Introduce plumbing for ObservablesReducer and its clients
In particular, some minor high level rearrangements are needed so that
the ObservablesReducer and step are always available before they are
potentially used e.g. in the LocalTopologyChecker in a future change.

Also, the local state and local topology are made earlier, so they
will be in place when needed for aspects of building the DD manager,
ie. LocalTopologyChecker. This reduces some minor code duplication,
also.

Minor unrelated clang-tidy fixes for forceelement.cpp

Refs #3887