WIP Add tests for orires
This is an attempt to add tests for orires. Currently it throws
at line 151 of orires.cpp due to conditions that I do not understand.
Once this condition, if it is not itself a bug, is satisfied, masses
will still need to be added before the test can actually not throw.

I changed an assert to a throw so that the test could actually run.