Separate preparing and distributing MDAtoms data to md algorithms
Now one function takes a pointer to MDAtoms and updates the data,
while another takes a const pointer and only hands out data. This
allows some simplification of branching relative to whether DD is
active because mdAlgorithmsDistributeAtomData is called whether
DD is active or not.

Further refactoring, in subsequent MRs, can lead to unification
of update_mdatoms with mdAlgorithmsPrepareAtomData.