Use ObservablesReducer for LINCS RMSD computation
Now the details of this computation are local to those concerned.

In particular, the RMSD components were only ever computed when
required, and are now also only reduced when required. The previous
logic reduced the RMSD components also when bCalcEner was true for
reasons other than the need for logging or energy output. This was
risky if there would be a large deviation and a long interval between
such output, as a floating-point exception could have occurred (albeit
very unlikely).

As a bonus, minimizers now can report the LINCS RMSD in their output.

Refs #3887 #3421