Introduce expanded ensemble in modular simulator
This introduces the expanded ensemble algorithm in modular simulator.
The implementation uses the legacy implementation to pick a new lambda
state, and communicates it to the FEP data element.

As in the legacy implementation, expanded ensemble is only implemented
for md-vv. Expanded ensemble is covered by the new free energy tests.

Refs #3417