Introduce simulated annealing for modular simulator
This introduces simulated annealing using a new element running
at the beginning of each step, and a new option for the reference
temperature manager.

Note that in the legacy implementation, only some of the algorithms
relying on the reference temperature update their behavior under
simulated annealing. Due to the small changes typical for each
annealing step and the fact that it is applied every step, this
is likely a reasonable choice. Either way, the current change does not
attempt to change this situation. In the spirit of the modular
simulator introduction, this merely replicates the legacy
behavior (and actually emphasizes it due to the well-defined
reference temperature update functionality). Discussing the behavior
of other algorithms under simulated annealing is beyond the scope
of this change, but is discussed at #4087.

Refs #3417