Full diff of nblib listed forces
This is not intended for detailed review, but to illustrate the
full scope of changes to nblib listed forces. Some changes are not
intended to be merged because work is ongoing to improve those
features, but they are included here to make it easy to see
everythin that could be changed.