Fix incorrect event dependency of GPU update
With separate PME ranks, staged communication and GPU update, the final
forces are produced on the CPU and re-uploaded to the GPU prior to
update. This upload was done on a "All" locality leading to an incorrect
dependency on a different (event corresponding to the GPU reduction). As
a result, incorrect forces could be used in GPU update.

This change moves this special case copy to the same locality used in
other force host to device copies to eliminate the bug.

Fixes #4130