Introduce nblib nonbonded force calculator impl class
Some data member structs are moved
to separate headers so that both cpu and (forthcoming) gpu force
calculators can depend on those structs. They are not tested directly
since they are not part of any interface, but rely on the integration
testing implicit in the fact that these structs are members of the
force calculator classes.