Do not wait for non-local coordinates D2H copy on search steps
The non local device to copy is not called on search steps, hence
the event should not be waited on these steps. they are not copied
on search steps. This mathes the copy call conditional with the
conditional around waiting. Note that the GPU buffer ops are never
used without PP DD, hence the check on the later is not needed around
wait for copy completion call.

Refs. #3988