Add lambda state to dhdl.xvg with AWH fep
Now the lambda state is printed to dhdl.xvg when using AWH fep,
like was already done for expanded ensemble. Introduced a boolean
for this, which simplifies the code and removes an argument.

Note that the lambda state is not yet printed to the free-energy
blocks in the energy file, as is done with expanded ensemble.
Due to the structure of the blocks, achieving that requires a lot
of changes.