Fix UB when generating local indices for constraints
Previously, for serial code, we were setting the anrs_loc to point to the
same memory as the global anrs. This caused UB when the local array was
reallocated but we were still reading from the global one.

Now, we properly signal that the array is empty, just like we do in
parallel code.

Issue introduced in MR !1904 (01b23435).

Closes #4177.