Optimize order of constraints in LINCS
Constraints were assigned on atom order. This works well when atoms
are ordered according to the topology, as the atoms and constraints
are usually ordered along chains. But with constraints=all-bonds and
domain decomposition, this leads to a large serial task when using
many OpenMP threads. Now connected constraints are assigned to the
same task as much as possible.

Refs #4195