Add second LINCS atom update task
Due to the change of atom ordering from topology to spatial grid
ordering in serial, the atom update tasks in LINCS have many more
dependencies. This leads to large performance degradation on many
threads, as the inter-dependent updates were done single-threaded.
To resolve this, as second set of update task is added. The gives
near perfect load balancing at the cost of an extra omp barrier.

Refs #4195