Support clang 13 in CI
Had to reinstate the ability for Docker containers to always
be based on the distro version that we require, after recent
refactoring.

Also changed to always use the llvm apt repository for installing
llvm, since it is clear that e.g. 20.04 has stopped adding support for
new llvm versions and we'll probably have this issue again with 22.04

Also updated supported compilers in install guide