Remove short-circuit logic for OpenMP setup
This is only called once during mdrun setup, and it is not a problem
to call it multiple times. (Thread safety is assured separately.)
It's necessary to call it multiple times, e.g. from test binaries or
API-driven simulations.

Having removed the short-circuit logic, there was no use for the
member variable that recorded whether initialization had taken place,
so it is removed.