Reimplement constant acceleration groups
This is a partial revert of be763d04. This removed the acceleration
group feature, because it allegedly was broken. The only issue was
that the reported group velocties were all zero. This issue is now
fixed by removing the acceleration group velocities from the output.
Some variable names are now more verbose than in the original code.

Fixes #1354
Refs #1971
Fixes #4251