Unify coordinate copy handling across GPU platforms
Instead of handling the OpenCL host to device copy as a special case
where this operation is issues in the PME stream, as the "Update" stream
is anyway created unconditionally it makes sense to simplify codepaths
at the negligible cost of an extra GPU-side synchronization.

This allows unifying the coordinate handling in stateGpu as well as in
PME spread (with combined PP+PME) removiung complexity and makes it
easier to reason about event consumption reducing the need for balancing
workarounds.

Refs #3988